# openapi/smartevents_mgmt.yaml
...
openapi: 3.0.3
info:
  title: Red Hat Openshift SmartEvents Fleet Manager
  description: The API exposed by the fleet manager of the SmartEvents service.
  contact:
    name: Development Team
    email: openbridge-dev@redhat.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 0.0.1
servers:
- url: https://api.stage.openshift.com
- url: http://localhost:8000
- url: https://api.openshift.com
- url: /
...
paths:
...
  /api/smartevents_mgmt/v1/errors/{id}:
    get:
      tags:
      - Error Catalog
      summary: Get an error from the error catalog.
      description: Get an error from the error catalog.
      operationId: ErrorsAPI_getError
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BridgeError'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorsList'
        "401":
          description: Unauthorized.
        "500":
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorsList'
...
  /api/smartevents_mgmt/v2/errors/{id}:
    get:
      tags:
      - Error Catalog
      summary: Get an error from the error catalog.
      description: Get an error from the error catalog.
      operationId: ErrorsAPIV2_getError
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BridgeError'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorsList'
        "401":
          description: Unauthorized.
        "500":
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorsList'
components:
  schemas:
...
    BridgeError:
      type: object
      properties:
        id:
          format: int32
          type: integer
        code:
          type: string
        reason:
          type: string
        type:
          $ref: '#/components/schemas/BridgeErrorType'
...
    BridgeErrorType:
      enum:
      - USER
      - PLATFORM
      type: string
...
    Error:
      required:
      - kind
      - reason
      - id
      - href
      type: object
      properties:
        kind:
          type: string
          nullable: false
        name:
          type: string
        code:
          type: string
        reason:
          type: string
          nullable: false
        id:
          type: string
          nullable: false
        href:
          type: string
          nullable: false
...
    ErrorsList:
      required:
      - kind
      type: object
      properties:
        kind:
          type: string
          nullable: false
        items:
          type: array
          items:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearer:
      type: http
      scheme: Bearer